% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\name{align,nmf-method}
\alias{align,nmf-method}
\alias{align.nmf}
\title{Align two NMF models}
\usage{
\S4method{align}{nmf}(object, ref, method = "cosine", ...)

align.nmf(w, wref, method = "cosine", ...)
}
\arguments{
\item{object}{nmf model to be aligned to \code{ref}}

\item{ref}{reference nmf model to which \code{object} will be aligned}

\item{method}{either \code{cosine} or \code{cor}}

\item{...}{arguments passed to or from other methods}

\item{w}{matrix with columns to be aligned to columns in \code{ref}}

\item{wref}{reference matrix to which columns in \code{w} will be aligned}
}
\description{
Align two NMF models
}
\details{
For \code{\link{nmf}} models, factors in \code{object} are reordered to minimize the cost of bipartite matching
(see \code{\link{bipartiteMatch}}) on a \code{\link{cosine}} or correlation distance matrix. The \eqn{w} matrix is used for matching, and
must be equidimensional in \code{object} and \code{ref}.
}
