% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\name{align}
\alias{align}
\alias{align,nmf-method}
\title{Align two NMF models}
\usage{
align(object, ...)

\S4method{align}{nmf}(object, ref, method = "cosine", ...)
}
\arguments{
\item{object}{nmf model to be aligned to \code{ref}}

\item{...}{arguments passed to or from other methods}

\item{ref}{reference nmf model to which \code{object} will be aligned}

\item{method}{either \code{cosine} or \code{cor}}
}
\value{
An \code{nmf} object with factors reordered to best match \code{ref}.
}
\description{
Align two NMF models
}
\details{
For \code{\link{nmf}} models, factors in \code{object} are reordered to minimize the cost of bipartite matching
(see \code{\link{bipartiteMatch}}) on a \code{\link{cosine}} or correlation distance matrix. The \eqn{w} matrix is used for matching, and
must be equidimensional in \code{object} and \code{ref}.
}
\examples{
\donttest{
data <- simulateNMF(50, 30, k = 3, seed = 1)
m1 <- nmf(data$A, 3, seed = 1, maxit = 50)
m2 <- nmf(data$A, 3, seed = 2, maxit = 50)
aligned <- align(m2, m1)  # reorder m2 factors to match m1
}
}
\seealso{
\code{\link{bipartiteMatch}}, \code{\link{cosine}}, \code{\link{nmf}}
}
