% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\name{align_models}
\alias{align_models}
\title{Align two matrices with bipartite matching}
\usage{
align_models(w, wref, method = "cosine", ...)
}
\arguments{
\item{w}{matrix with columns to be aligned to columns in \code{wref}}

\item{wref}{reference matrix to which columns in \code{w} will be aligned}

\item{...}{additional arguments}
}
\description{
Same as the \code{align} S4 method for the `nmf` class, but operates only on the `w` matrices.
}
