% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{as.data.frame.nmf_assessment}
\alias{as.data.frame.nmf_assessment}
\title{Convert assessment results to a one-row data frame}
\usage{
\method{as.data.frame}{nmf_assessment}(x, ...)
}
\arguments{
\item{x}{An \code{nmf_assessment} object.}

\item{...}{Ignored.}
}
\value{
A one-row data frame with all metrics as columns.
}
\description{
Convert assessment results to a one-row data frame
}
