% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_log.R
\name{as.data.frame.training_logger}
\alias{as.data.frame.training_logger}
\title{Convert training log to data.frame}
\usage{
\method{as.data.frame}{training_logger}(x, ...)
}
\arguments{
\item{x}{A \code{training_logger} object.}

\item{...}{Ignored.}
}
\value{
A data.frame with one row per logged iteration.
}
\description{
Convert training log to data.frame
}
