% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark structure performance}
\usage{
benchmark(x, y, n_reps = 100L)
}
\arguments{
\item{x}{sparse matrix object of class \code{dgCMatrix} of dimensions \code{m x n} with integral non-zero values}

\item{y}{dense matrix object of class \code{matrix} of dimensions \code{n x k}}

\item{n_reps}{number of reps}
}
\value{
global variables \code{bench_colsums} and \code{bench_tcrossprod} in the global environment. Use \code{RcppClock::plot} method to visualize results, or coerce to a data.frame.
}
\description{
Measure runtime for computing column sums and sparse-dense matrix multiplication
}
\examples{
library(Matrix)
x <- rsparsematrix(50, 100, density = 0.5,
  rand.x = function(n) { sample(1:10, n, replace = TRUE)})
y <- matrix(runif(10 * 100), 10, 100)
VSE::benchmark(x, y)
plot(bench_tcrossprod)
plot(bench_colsums)
}
\seealso{
\code{\link{tcrossprod}}
}
