% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_plots.R
\name{biplot.nmf}
\alias{biplot.nmf}
\title{Biplot for NMF factors}
\usage{
\method{biplot}{nmf}(x, factors = c(1, 2), matrix = "w", group_by = NULL, ...)
}
\arguments{
\item{x}{an object of class "\code{nmf}"}

\item{factors}{length 2 vector specifying factors to plot.}

\item{matrix}{either \code{w} or \code{h}}

\item{group_by}{a discrete factor giving groupings for samples or features. Must be of the same length as number of samples in \code{object$h} or number of features in \code{object$w}.}

\item{...}{for consistency with \code{biplot} generic}
}
\value{
ggplot2 object
}
\description{
Produces a biplot from the output of \code{\link{nmf}}
}
\seealso{
\code{\link{nmf}}
}
