% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BloodCells.R
\docType{data}
\name{BloodCells}
\alias{BloodCells}
\title{Interferon-stimulated blood cells}
\format{
sparse \code{Matrix::dgCMatrix} of gene transcript counts (rows) vs. single cells (columns), labeled as either stimulated or not stimulated (colnames).
}
\usage{
data(BloodCells)
}
\description{
Transcript counts for 2000 variable genes in 1000 peripheral blood mononuclear cells (PBMCs), roughly half of them stimulated with IFNB
}
\details{
This dataset was derived from https://www.nature.com/articles/nbt.4042 and SeuratData.
}
\keyword{datasets}
