% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{calc_redundancy}
\alias{calc_redundancy}
\title{Calculate redundancy of data per column}
\usage{
calc_redundancy(A)
}
\arguments{
\item{A}{sparse matrix of class \code{dgCMatrix}}
}
\value{
fractional redundancy by column
}
\description{
Calculate redundancy of data per column
}
