% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colsums}
\alias{colsums}
\title{Compute column sums of sparse-encoded data}
\usage{
colsums(A, encoding = "CSC")
}
\arguments{
\item{A}{object of class \code{dgCMatrix}}

\item{encoding}{sparse encoding to be used}
}
\value{
vector of column sums
}
\description{
Compute column sums of sparse-encoded data
}
\examples{
library(Matrix)
A <- rsparsematrix(100, 1000, density = 0.5,
  rand.x = function(n) { sample(1:10, n, replace = TRUE)})
v <- colsums(A, encoding = "TRCSC_NP")
plot(v, Matrix::colSums(A))
}
