% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cumcm2011a}
\alias{cumcm2011a}
\title{Location of 319 sample site.}
\format{A data frame with 319 rows and 12 variables:
\describe{
  \item{x}{x coordinate, in m}
  \item{y}{y coordinate, in m}
  \item{z}{z coordinate, in m}
  \item{As}{As, Arsenic, in ug/g}
  \item{Cd}{y coordinate, in m}
  \item{Cr}{z coordinate, in m}
  \item{Cu}{x coordinate, in m}
  \item{Hg}{y coordinate, in m}
  \item{Ni}{z coordinate, in m}
  \item{Pb}{x coordinate, in m}
  \item{Zn}{y coordinate, in m}
  \item{area}{z coordinate, in m}
  ...
}}
\source{
\url{http://www.mcm.edu.cn/}
}
\usage{
cumcm2011a
}
\description{
A dataset containing the coordinate and other attributes of almost 319 sample
}
\keyword{datasets}
