% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclust.R
\name{.dclust_to_dendro}
\alias{.dclust_to_dendro}
\title{Convert dclust result to dendrogram}
\usage{
.dclust_to_dendro(clusters)
}
\value{
A \code{\link{dendrogram}} object.
}
\description{
Convert dclust result to dendrogram
}
\keyword{internal}
