% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_backend.R
\name{.find_gpu_lib}
\alias{.find_gpu_lib}
\title{Find GPU shared library}
\usage{
.find_gpu_lib()
}
\value{
File path to the GPU shared library, or \code{NULL}.
}
\description{
Find GPU shared library
}
\keyword{internal}
