% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{.fn_build_descriptor}
\alias{.fn_build_descriptor}
\title{Build a flat descriptor for Rcpp_factor_net_fit()}
\usage{
.fn_build_descriptor(net)
}
\description{
Build a flat descriptor for Rcpp_factor_net_fit()
}
\keyword{internal}
