% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{.fn_find_layer_idx}
\alias{.fn_find_layer_idx}
\title{Find the index of a node in the layers list}
\usage{
.fn_find_layer_idx(node, layers)
}
\description{
Find the index of a node in the layers list
}
\keyword{internal}
