% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{.fn_has_r_callbacks}
\alias{.fn_has_r_callbacks}
\title{Check if any layer uses R-side callback or reference guides}
\usage{
.fn_has_r_callbacks(layers)
}
\description{
Check if any layer uses R-side callback or reference guides
}
\keyword{internal}
