% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{.fn_wrap_graph_result}
\alias{.fn_wrap_graph_result}
\title{Wrap C++ graph result into factor_net_result}
\usage{
.fn_wrap_graph_result(raw, net)
}
\value{
A \code{factor_net_result} object.
}
\description{
Wrap C++ graph result into factor_net_result
}
\keyword{internal}
