% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_validation.R
\name{.load_h5_sparse_group}
\alias{.load_h5_sparse_group}
\title{Load sparse matrix from HDF5 group (10x-style CSC)}
\usage{
.load_h5_sparse_group(grp)
}
\arguments{
\item{grp}{An hdf5r H5Group with data/indices/indptr/shape}
}
\value{
A dgCMatrix
}
\description{
Load sparse matrix from HDF5 group (10x-style CSC)
}
\keyword{internal}
