% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dots.R
\name{.parse_nmf_dots}
\alias{.parse_nmf_dots}
\title{Parse advanced NMF parameters from dots}
\usage{
.parse_nmf_dots(dots)
}
\arguments{
\item{dots}{list from list(...)}
}
\value{
Named list with all advanced parameters filled to defaults
}
\description{
Parse advanced NMF parameters from dots
}
\keyword{internal}
