% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclust.R
\name{.plot_dclust_composition}
\alias{.plot_dclust_composition}
\title{Plot composition bars aligned with dendrogram leaves}
\usage{
.plot_dclust_composition(clusters, dendro, labels, palette)
}
\value{
Called for its side effect (plotting). Returns \code{NULL} invisibly.
}
\description{
Plot composition bars aligned with dendrogram leaves
}
\keyword{internal}
