% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclust.R
\name{.try_gpu_dclust}
\alias{.try_gpu_dclust}
\title{Try GPU dispatch for divisive clustering}
\usage{
.try_gpu_dclust(A, min_samples, min_dist, tol, maxit, nonneg, seed, verbose)
}
\description{
Try GPU dispatch for divisive clustering
}
\keyword{internal}
