% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_log.R
\name{export_log}
\alias{export_log}
\title{Export training log to CSV}
\usage{
export_log(logger, file)
}
\arguments{
\item{logger}{A \code{training_logger} object.}

\item{file}{Path to write the CSV file.}
}
\value{
Invisibly returns the data frame written.
}
\description{
Export training log to CSV
}
\seealso{
\code{\link{training_logger}}, \code{\link{as.data.frame.training_logger}}
}
