% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_net.R
\name{factor_concat}
\alias{factor_concat}
\title{Concatenate H factors from branches (row-bind)}
\usage{
factor_concat(...)
}
\arguments{
\item{...}{Two or more \code{fn_node} objects (layer outputs).}
}
\value{
An \code{fn_node} of type "concat".
}
\description{
Creates a node that row-binds the H factors from multiple branches.
Allows combining branches with different ranks into a single
higher-dimensional representation: k_out = k_1 + k_2 + ...
}
\examples{
data(aml)
inp <- factor_input(aml)
branch1 <- nmf_layer(inp, k = 3)
branch2 <- nmf_layer(inp, k = 5)
combined <- factor_concat(branch1, branch2)
}
