% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastnmf}
\alias{fastnmf}
\title{Fast NMF}
\usage{
fastnmf(data, w, tol = 1e-04, maxit = 100L)
}
\arguments{
\item{w}{dense matrix of factors in rows and features in columns}

\item{tol}{tolerance of the fit}

\item{maxit}{maximum number of fitting iterations}

\item{sparse}{matrix of features in rows and samples in columns, of class \code{dgCMatrix}}
}
\description{
A basic implementation of NMF
}
