% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{getRcppMLthreads}
\alias{getRcppMLthreads}
\title{Get the number of threads RcppML should use}
\usage{
getRcppMLthreads()
}
\value{
integer giving number of threads to be used by RcppML functions. \code{0} corresponds to all available threads, as determined by OpenMP.
}
\description{
Get the number of threads that will be used by RcppML functions supporting parallelization with OpenMP. Use \code{\link{setRcppMLthreads}} to set the number of threads to be used.
}
\examples{
# set parallel configuration, 
# letting OpenMP decide how many threads to use
setRcppMLthreads()
getRcppMLthreads()

# set serial configuration
setRcppMLthreads(1)
getRcppMLthreads()

}
\seealso{
\code{\link{setRcppMLthreads}}
}
\author{
Zach DeBruine
}
