% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_backend.R
\name{gpu_info}
\alias{gpu_info}
\title{Get GPU device information}
\usage{
gpu_info()
}
\value{
data.frame with GPU device details, or NULL if no GPU
}
\description{
Get GPU device information
}
\examples{
gpu_info()
}
\seealso{
\code{\link{gpu_available}}, \code{\link{nmf}}
}
