% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_gpu.R
\name{gpu_sparse_matrix-methods}
\alias{gpu_sparse_matrix-methods}
\alias{print.gpu_sparse_matrix}
\alias{dim.gpu_sparse_matrix}
\alias{nrow.gpu_sparse_matrix}
\alias{ncol.gpu_sparse_matrix}
\title{Methods for gpu_sparse_matrix objects}
\usage{
\method{print}{gpu_sparse_matrix}(x, ...)

\method{dim}{gpu_sparse_matrix}(x)

\method{nrow}{gpu_sparse_matrix}(x)

\method{ncol}{gpu_sparse_matrix}(x)
}
\arguments{
\item{x}{a \code{gpu_sparse_matrix} object}

\item{...}{additional arguments (unused)}
}
\value{
\describe{
  \item{\code{print}}{Invisibly returns \code{x}, prints summary to console.}
  \item{\code{dim}}{Integer vector of length 2: \code{c(nrow, ncol)}.}
  \item{\code{nrow}}{Number of rows (integer).}
  \item{\code{ncol}}{Number of columns (integer).}
}
}
\description{
S3 methods for the \code{gpu_sparse_matrix} class returned by
\code{\link{st_read_gpu}}.
}
\examples{
\dontrun{
gpu_mat <- st_read_gpu("data.spz")
print(gpu_mat)
dim(gpu_mat)
nrow(gpu_mat)
ncol(gpu_mat)
st_free_gpu(gpu_mat)
}
}
\seealso{
\code{\link{st_read_gpu}}, \code{\link{st_free_gpu}}
}
