% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{View internal structure of sparse encodings}
\usage{
inspect(A, encoding = "TRCSC_NP")
}
\arguments{
\item{A}{object of class \code{dgCMatrix}}

\item{encoding}{sparse encoding to be used}
}
\value{
debug level output to console
}
\description{
View internal structure of sparse encodings
}
\examples{
library(Matrix)
A <- rsparsematrix(20, 5, density = 0.2,
  rand.x = function(n) { sample(1:5, n, replace = TRUE)})
inspect(A, "TRCSC_NP")
}
