% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnmf.R, R/nmf_methods.R
\docType{class}
\name{lnmf}
\alias{lnmf}
\alias{lnmf,}
\alias{lnmf-class}
\title{Linked non-negative matrix factorization}
\usage{
lnmf(
  data,
  k_wh,
  k_uv,
  tol = 1e-04,
  maxit = 100,
  L1 = c(0, 0),
  L2 = c(0, 0),
  nonneg = TRUE,
  seed = NULL,
  mask = NULL
)
}
\arguments{
\item{data}{list of dense or sparse matrices giving features in rows and samples in columns. Rows in all matrices must correspond exactly. Prefer \code{matrix} or \code{Matrix::dgCMatrix}, respectively}

\item{k_wh}{rank of the shared factorization}

\item{k_uv}{ranks of the unique factorizations, an array corresponding to each dataset provided}

\item{tol}{tolerance of the fit}

\item{maxit}{maximum number of fitting iterations}

\item{L1}{LASSO penalties in the range (0, 1], single value or array of length two for \code{c(w & u, h & v)}}

\item{L2}{Ridge penalties greater than zero, single value or array of length two for \code{c(w & u, h & v)}}

\item{nonneg}{enforce non-negativity}

\item{seed}{single initialization seed or array of seeds. If multiple seeds are provided, the model with least mean squared error is returned.}

\item{mask}{list of dense or sparse matrices indicating values in \code{data} to handle as missing. Prefer \code{Matrix::ngCMatrix}. Alternatively, specify a string "\code{zeros}" or "\code{NA}" to mask either all zeros or NA values.}
}
\value{
an object of class \code{lnmf}
}
\description{
Run lNMF on a list of datasets to separate shared and unique signals.
}
\details{
Detailed documentation to come
}
\section{Slots}{

\describe{
\item{\code{w}}{feature factor matrix of shared signal across all datasets}

\item{\code{d_wh}}{scaling diagonal vector for \code{w} and each \code{h} matrix}

\item{\code{h}}{sample factor matrix mapping shared signal from \code{w} to each dataset}

\item{\code{u}}{feature factor matrix of unique signal for each dataset}

\item{\code{d_uv}}{scaling diagonal vector for \code{u} and \code{v} for each dataset}

\item{\code{v}}{sample factor matrix mapping unique signal from \code{u} to each dataset}

\item{\code{misc}}{list often containing components:
\itemize{
  \item tol     : tolerance of fit
  \item iter    : number of fitting updates
  \item runtime : runtime in seconds
  \item mse     : mean squared error of model (calculated for multiple starts only)
  \item w_init  : initial w matrix used for model fitting
}}
}}

\references{
DeBruine, ZJ, Melcher, K, and Triche, TJ. (2021). "High-performance non-negative matrix factorization for large single-cell data." BioRXiv.
}
\seealso{
\code{\link{nmf}}, \code{\link{predict.nmf}}, \code{\link{evaluate.nmf}}, \code{\link{nnls}}
}
\author{
Zach DeBruine
}
