% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{memuse}
\alias{memuse}
\title{Get memory used by C++ object}
\usage{
memuse(A, encoding = "CSC")
}
\arguments{
\item{A}{object of class \code{dgCMatrix}}

\item{encoding}{sparse encoding to be used}
}
\value{
size of object in bytes
}
\description{
Calculates size of vector container, all sub-containers, and all index and value types within the container.  This function stores all values and indices as \code{short int}.
}
\examples{
library(Matrix)
A <- rsparsematrix(100, 1000, density = 0.5,
  rand.x = function(n) { sample(1:10, n, replace = TRUE)})
memuse(A, "TRCSC_NP")
}
