% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movielens}
\alias{movielens}
\title{MovieLens Dataset}
\format{
A \code{dgCMatrix} sparse matrix (3,867 x 610) containing movie ratings.
  Rows are movies, columns are users, and values are ratings from 1-5 stars.
  Most entries are zero (unrated). Genre information is stored as an attribute.

  Access genre data via \code{attr(movielens, "genres")}, which returns an
  \code{ngCMatrix} sparse binary matrix (19 x 3,867) indicating genre membership.
  Rows are genres, columns are movies, and entries are TRUE/FALSE for genre membership.
}
\source{
MovieLens 100K Dataset from GroupLens Research.
\url{https://grouplens.org/datasets/movielens/}
}
\usage{
movielens
}
\description{
Movie ratings data from the MovieLens 100K dataset, containing user ratings
for movies along with genre information.
}
\details{
This is a subset of the MovieLens dataset, widely used for demonstrating
collaborative filtering and recommendation systems. The data is highly sparse,
as most users have only rated a small fraction of available movies.

The 19 genres include:
\itemize{
  \item Action, Adventure, Animation, Children
  \item Comedy, Crime, Documentary, Drama
  \item Fantasy, Film-Noir, Horror, Musical
  \item Mystery, Romance, Sci-Fi, Thriller
  \item War, Western, IMAX
}
}
\examples{
\donttest{
# Load dataset
data(movielens)

# Inspect structure
dim(movielens)
dim(attr(movielens, "genres"))

# Sparsity
Matrix::nnzero(movielens) / prod(dim(movielens))

# Run NMF for collaborative filtering
result <- nmf(movielens, k = 20, maxit = 50)
}

}
\keyword{datasets}
