% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\name{new_nmf}
\alias{new_nmf}
\title{Constructor for nmf class}
\usage{
new_nmf(w, d, h, tol = NA_real_, iter = NA_real_, runtime = NA_real_)
}
\arguments{
\item{w}{matrix of features (rows) by factors (columns)}

\item{d}{scaling diagonal vector for factors}

\item{h}{matrix of factors (rows) by samples (columns)}

\item{tol}{if applicable, tolerance of the fit}

\item{iter}{if applicable, number of iterations for fitting}

\item{runtime}{if applicable, runtime of the fit}
}
\description{
Construct and validate an S3 \code{nmf} object.
}
