% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\docType{class}
\name{nmf-class}
\alias{nmf-class}
\title{nmf S4 Class}
\description{
The S4 class for NMF model results.
}
\section{Slots}{

\describe{
\item{\code{w}}{feature factor matrix}

\item{\code{d}}{scaling diagonal vector}

\item{\code{h}}{sample factor matrix}

\item{\code{misc}}{list containing optional components including tol (tolerance), iter (iterations), 
loss (final loss value), loss_type (loss function used), runtime (in seconds), 
w_init (initial w matrix), test_mask (CV test set), test_seed (CV seed), 
test_fraction (CV holdout fraction), train_loss (CV training loss), 
test_loss (CV test loss), and best_iter (CV best iteration)}
}}

\seealso{
\code{\link{nmf}}, \code{\link{evaluate}}, \code{\link{align}}, \code{\link{predict,nmf-method}}
}
