% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\docType{methods}
\name{subset,nmf-method}
\alias{subset,nmf-method}
\alias{[,nmf,ANY,ANY,ANY-method}
\alias{head,nmf-method}
\alias{show,nmf-method}
\alias{dimnames,nmf-method}
\alias{dim,nmf-method}
\alias{t,nmf-method}
\alias{sort,nmf-method}
\alias{prod,nmf-method}
\alias{$,nmf-method}
\alias{coerce,nmf,list-method}
\alias{coerce,lnmf,nmf-method}
\alias{[[,lnmf-method}
\alias{head,lnmf-method}
\alias{show,lnmf-method}
\title{nmf class methods}
\usage{
\S4method{subset}{nmf}(x, i, ...)

\S4method{[}{nmf,ANY,ANY,ANY}(x, i)

\S4method{head}{nmf}(x, n = getOption("digits"), ...)

\S4method{show}{nmf}(object)

\S4method{dimnames}{nmf}(x)

\S4method{dim}{nmf}(x)

\S4method{t}{nmf}(x)

\S4method{sort}{nmf}(x, decreasing = TRUE, ...)

\S4method{prod}{nmf}(x, ..., na.rm = FALSE)

\S4method{$}{nmf}(x, name)

\S4method{coerce}{nmf,list}(from, to)

\S4method{coerce}{lnmf,nmf}(from, to)

\S4method{[[}{lnmf}(x, i)

\S4method{head}{lnmf}(x, n = getOption("digits"), ...)

\S4method{show}{lnmf}(object)
}
\description{
nmf class methods
}
