% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf2.R
\name{nmf2}
\alias{nmf2}
\title{Rank-2 non-negative matrix factorization}
\usage{
nmf2(
  A,
  tol = 1e-04,
  maxit = 100,
  verbose = TRUE,
  nonneg = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{A}{matrix of features x samples, may be dense or sparse (a class of the "Matrix" package coercible to \code{Matrix::dgCMatrix})}

\item{tol}{correlation distance between \eqn{w} across consecutive iterations at which to stop factorization}

\item{maxit}{maximum number of alternating updates of \eqn{w} and \eqn{h}}

\item{verbose}{print tolerances after each iteration to the console}

\item{nonneg}{apply non-negativity constraints}

\item{seed}{random seed for initializing \eqn{h}.}

\item{...}{advanced parameters, see details}
}
\value{
A list giving the factorization model:
\itemize{
\item w    : feature factor matrix
\item d    : scaling diagonal vector
\item h    : sample factor matrix
\item tol  : tolerance between models after final update
\item iter : number of alternating iterations completed
}
}
\description{
High-performance rank-2 matrix factorization with optional non-negativity constraints.
}
\details{
here
}
\section{Advanced parameters}{

Several parameters hidden in the \code{...} argument may be adjusted (although defaults should entirely satisfy) in addition to those documented explicitly:
\itemize{
\item \code{diag}, default TRUE. Enable model diagonalization to normalize factors to sum to 1, guarantee symmetry for symmetric factorizations, and distribute L1 penalties consistently and equally across all factors.
\item \code{samples}, samples (from 1 to \code{ncol(A)}) in \code{A} to include in factorization. Default \code{1:ncol(A)}.
}
}

\seealso{
\code{\link{nmf}}, \code{\link{bipartition}}
}
\author{
Zach DeBruine
}
