% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmf.R
\name{plot_nmf_convergence}
\alias{plot_nmf_convergence}
\title{Plot Convergence Analysis}
\usage{
plot_nmf_convergence(x, smooth, span, interactive, theme)
}
\value{
A \code{ggplot2} object (or \code{plotly} if \code{interactive = TRUE})
}
\description{
Plot Convergence Analysis
}
\keyword{internal}
