% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmf.R
\name{plot_nmf_loss}
\alias{plot_nmf_loss}
\title{Plot Loss History}
\usage{
plot_nmf_loss(x, smooth, span, log_scale, interactive, theme)
}
\value{
A \code{ggplot2} object (or \code{plotly} if \code{interactive = TRUE})
}
\description{
Plot Loss History
}
\keyword{internal}
