% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmf.R
\name{plot_nmf_regularization}
\alias{plot_nmf_regularization}
\title{Plot Regularization Breakdown}
\usage{
plot_nmf_regularization(x, smooth, span, interactive, theme)
}
\value{
A \code{ggplot2} object showing regularization contributions
}
\description{
Plot Regularization Breakdown
}
\keyword{internal}
