% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmf.R
\name{plot_nmf_sparsity}
\alias{plot_nmf_sparsity}
\title{Plot Factor Sparsity}
\usage{
plot_nmf_sparsity(x, interactive, theme)
}
\value{
A \code{ggplot2} object showing sparsity of W and H matrices
}
\description{
Plot Factor Sparsity
}
\keyword{internal}
