% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_net.R
\name{print.factor_net}
\alias{print.factor_net}
\title{Print a factor_net}
\usage{
\method{print}{factor_net}(x, ...)
}
\arguments{
\item{x}{A \code{factor_net} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a factor_net
}
