% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate_graph.R
\name{print.factor_net_cv}
\alias{print.factor_net_cv}
\title{Print a factor_net_cv result}
\usage{
\method{print}{factor_net_cv}(x, ...)
}
\arguments{
\item{x}{A \code{factor_net_cv} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a factor_net_cv result
}
