% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_methods.R
\name{print.factor_net_result}
\alias{print.factor_net_result}
\title{Print a factor_net_result}
\usage{
\method{print}{factor_net_result}(x, ...)
}
\arguments{
\item{x}{A \code{factor_net_result} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a factor_net_result
}
