% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_net.R
\name{print.fn_factor_config}
\alias{print.fn_factor_config}
\title{Print an fn_factor_config}
\usage{
\method{print}{fn_factor_config}(x, ...)
}
\arguments{
\item{x}{An \code{fn_factor_config} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an fn_factor_config
}
