% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_net.R
\name{print.fn_node}
\alias{print.fn_node}
\title{Print an fn_node}
\usage{
\method{print}{fn_node}(x, ...)
}
\arguments{
\item{x}{An \code{fn_node} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an fn_node
}
\seealso{
\code{\link{factor_input}}, \code{\link{nmf_layer}}, \code{\link{factor_net}}
}
