% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{print.nmf_assessment}
\alias{print.nmf_assessment}
\title{Print method for nmf_assessment objects}
\usage{
\method{print}{nmf_assessment}(x, ...)
}
\arguments{
\item{x}{an \code{nmf_assessment} object}

\item{...}{additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print method for nmf_assessment objects
}
