% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.R
\name{print.nmf_guide}
\alias{print.nmf_guide}
\title{Print an nmf_guide}
\usage{
\method{print}{nmf_guide}(x, ...)
}
\arguments{
\item{x}{An \code{nmf_guide} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an nmf_guide
}
