% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_log.R
\name{print.training_logger}
\alias{print.training_logger}
\title{Print a training log}
\usage{
\method{print}{training_logger}(x, ...)
}
\arguments{
\item{x}{A \code{training_logger} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a training log
}
\seealso{
\code{\link{training_logger}}
}
