% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_get_available_ram_mb}
\alias{Rcpp_get_available_ram_mb}
\title{Get available system RAM in megabytes}
\usage{
Rcpp_get_available_ram_mb()
}
\value{
Double: available RAM in MB
}
\description{
Get available system RAM in megabytes
}
\keyword{internal}
