% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_sp_compress}
\alias{Rcpp_sp_compress}
\title{Convert a dgCMatrix to SparsePress format (in-memory)}
\usage{
Rcpp_sp_compress(A, use_delta = TRUE, use_value_pred = FALSE)
}
\arguments{
\item{A}{A sparse matrix (dgCMatrix)}

\item{use_delta}{Use delta prediction}

\item{use_value_pred}{Use value prediction}
}
\value{
A raw vector containing the compressed data
}
\description{
Convert a dgCMatrix to SparsePress format (in-memory)
}
\keyword{internal}
