% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_sp_metadata}
\alias{Rcpp_sp_metadata}
\title{Get metadata from a SparsePress (.spz) file without decompressing}
\usage{
Rcpp_sp_metadata(path)
}
\arguments{
\item{path}{Input file path (.spz)}
}
\value{
A list with matrix metadata
}
\description{
Get metadata from a SparsePress (.spz) file without decompressing
}
\keyword{internal}
