% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_sp_read_transpose}
\alias{Rcpp_sp_read_transpose}
\title{Read the pre-stored transpose from a v2 SparsePress (.spz) file}
\usage{
Rcpp_sp_read_transpose(path)
}
\arguments{
\item{path}{Input file path (.spz)}
}
\value{
A sparse matrix (dgCMatrix) containing CSC(A^T)
}
\description{
Read the pre-stored transpose from a v2 SparsePress (.spz) file
}
\keyword{internal}
