% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_st_add_transpose}
\alias{Rcpp_st_add_transpose}
\title{Add transpose section to an existing v2 .spz file}
\usage{
Rcpp_st_add_transpose(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to the .spz v2 file}

\item{verbose}{Logical; print progress}
}
\value{
Logical TRUE on success
}
\description{
Add transpose section to an existing v2 .spz file
}
\keyword{internal}
