% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_st_read_obs}
\alias{Rcpp_st_read_obs}
\title{Read obs table from a v2 .spz file}
\usage{
Rcpp_st_read_obs(path)
}
\arguments{
\item{path}{Path to .spz file}
}
\value{
A data.frame, or empty data.frame if no obs table
}
\description{
Read obs table from a v2 .spz file
}
\keyword{internal}
