% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_st_read_var}
\alias{Rcpp_st_read_var}
\title{Read var table from a v2 .spz file}
\usage{
Rcpp_st_read_var(path)
}
\arguments{
\item{path}{Path to .spz file}
}
\value{
A data.frame, or empty data.frame if no var table
}
\description{
Read var table from a v2 .spz file
}
\keyword{internal}
