% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_st_serialize_table}
\alias{Rcpp_st_serialize_table}
\title{Serialize a data.frame to obs/var table binary format}
\usage{
Rcpp_st_serialize_table(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A raw vector containing the serialized table
}
\description{
Serialize a data.frame to obs/var table binary format
}
\keyword{internal}
