% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{RcppML-deprecated}
\alias{RcppML-deprecated}
\title{Deprecated functions from RcppML 0.3.7}
\description{
These functions are provided for backward compatibility with the CRAN 0.3.7
API but are deprecated and will be removed in a future version.
}
\keyword{internal}
