% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppML-package.R
\docType{package}
\name{RcppML}
\alias{RcppML}
\alias{_PACKAGE}
\alias{RcppML-package}
\title{RcppML: Machine Learning in Rcpp}
\description{
Constrained and regularized least squares and matrix factorization model projection for dense and sparse matrices.
}
\details{
A library of machine learning methods in Rcpp being actively developed using a very fast active-set/coordinate descent constrained least squares algorithm with optional L0 regularization.
}
\author{
Zach DeBruine
}
