% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve a linear system with coordinate descent}
\usage{
solve(
  a,
  b,
  cd_maxit = 100L,
  cd_tol = 1e-08,
  L1 = 0,
  L2 = 0,
  upper_bound = 0,
  nonneg = TRUE
)
}
\arguments{
\item{a}{symmetric positive definite matrix giving coefficients of the linear system}

\item{b}{matrix giving the right-hand side(s) of the linear system}

\item{cd_maxit}{maximum number of coordinate descent iterations}

\item{cd_tol}{stopping criteria, difference in \eqn{x} across consecutive solutions over the sum of \eqn{x}}

\item{L1}{L1/LASSO penalty to be subtracted from \code{b}}

\item{L2}{Ridge penalty to be added to the diagonal of \code{a}}

\item{upper_bound}{maximum value permitted in solution, set to \code{0} to impose no upper bound}

\item{nonneg}{if TRUE (default), impose non-negativity constraint on solution}
}
\value{
vector or matrix giving solution for \code{x}
}
\description{
Solves the equation \code{a \%*\% x = b} for \code{x}, optionally subject to \eqn{x >= 0}.
}
\details{
This is a very fast implementation of sequential coordinate descent least squares, suitable for very small or very large systems.
The algorithm begins with a zero-filled initialization of \code{x}.

Least squares by \strong{sequential coordinate descent} is used to ensure the solution returned is exact. This algorithm was
introduced by Franc et al. (2005), and our implementation is a vectorized and optimized rendition of that found in the NNLM R package by Xihui Lin (2020).

When \code{nonneg = TRUE}, this solves the Non-Negative Least Squares (NNLS) problem.
When \code{nonneg = FALSE}, this solves unconstrained least squares.
}
\note{
This function is deprecated. Use \code{\link{nnls}} instead.
\code{solve()} previously masked \code{base::solve()}.
}
\examples{
\dontrun{
# compare solution to base::solve for a random system
X <- matrix(runif(100), 10, 10)
a <- crossprod(X)
b <- crossprod(X, runif(10))

# unconstrained solution (same as base::solve)
unconstrained_soln <- solve(a, b, nonneg = FALSE)

# non-negative constrained solution
nonneg_soln <- solve(a, b, nonneg = TRUE)
}
}
\references{
DeBruine, ZJ, Melcher, K, and Triche, TJ. (2021). "High-performance non-negative matrix factorization for large single-cell data." BioRXiv.

Franc, VC, Hlavac, VC, and Navara, M. (2005). "Sequential Coordinate-Wise Algorithm for the Non-negative Least Squares Problem. Proc. Int'l Conf. Computer Analysis of Images and Patterns."

Lin, X, and Boutros, PC (2020). "Optimization and expansion of non-negative matrix factorization." BMC Bioinformatics.
}
\seealso{
\code{\link{nmf}}, \code{\link{nnls}}
}
\author{
Zach DeBruine
}
\keyword{internal}
