% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsepress.R
\name{sp_compress}
\alias{sp_compress}
\alias{sp_decompress}
\title{Convert a sparse matrix to/from SparsePress format in memory}
\usage{
sp_compress(x, delta = TRUE, value_pred = FALSE)

sp_decompress(x)
}
\arguments{
\item{x}{For \code{sp_compress}: a \code{dgCMatrix}. For
\code{sp_decompress}: a raw vector from \code{sp_compress}.}

\item{delta}{Logical; use delta prediction. Default \code{TRUE}.}

\item{value_pred}{Logical; use value prediction. Default \code{FALSE}.}
}
\value{
\code{sp_compress} returns a raw vector. \code{sp_decompress}
  returns a \code{dgCMatrix}.
}
\description{
Convert a sparse matrix to/from SparsePress format in memory
}
\examples{
\dontrun{
library(Matrix)
A <- rsparsematrix(100, 50, 0.1)
blob <- sp_compress(A)
B <- sp_decompress(blob)
all.equal(A, B)
}

}
\seealso{
\code{\link{sp_read}}, \code{\link{sp_write}}
}
\keyword{internal}
