% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsepress.R
\name{sparsepress-deprecated}
\alias{sparsepress-deprecated}
\alias{sp_write}
\alias{sp_read}
\alias{sp_read_transpose}
\alias{sp_info}
\alias{sp_compress}
\alias{sp_decompress}
\alias{sp_write_dense}
\alias{sp_read_dense}
\alias{sp_convert}
\title{SparsePress I/O (Deprecated)}
\usage{
sp_write(
  x,
  path,
  delta = TRUE,
  value_pred = FALSE,
  verbose = FALSE,
  precision = "auto",
  row_sort = FALSE,
  include_transpose = FALSE,
  chunk_cols = 2048L
)

sp_read(path, cols = NULL, reorder = TRUE)

sp_read_transpose(path)

sp_info(path)

sp_compress(x, delta = TRUE, value_pred = FALSE)

sp_decompress(x)

sp_write_dense(
  x,
  path,
  include_transpose = FALSE,
  chunk_cols = 2048L,
  codec = "raw",
  delta = FALSE,
  verbose = FALSE
)

sp_read_dense(path)

sp_convert(
  input,
  output,
  precision = "auto",
  include_transpose = FALSE,
  row_sort = TRUE,
  verbose = 1L
)
}
\description{
These functions are deprecated. Use the \code{st_*} equivalents from
the \code{\link{streampress}} API instead.
}
\seealso{
\code{\link{streampress}}, \code{\link{st_write}}, \code{\link{st_read}},
  \code{\link{st_info}}, \code{\link{st_write_dense}}, \code{\link{st_read_dense}}
}
\keyword{internal}
