% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streampress.R
\name{st_add_transpose}
\alias{st_add_transpose}
\title{Add Transpose Section to an Existing StreamPress File}
\usage{
st_add_transpose(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to a \code{.spz} file.}

\item{verbose}{Logical; print progress. Default TRUE.}
}
\value{
Invisibly returns the path.
}
\description{
Add Transpose Section to an Existing StreamPress File
}
\examples{
\donttest{
library(Matrix)
m <- rsparsematrix(50, 20, 0.3)
tmp <- tempfile(fileext = ".spz")
st_write(m, tmp, include_transpose = FALSE)
st_add_transpose(tmp)
info <- st_info(tmp)
unlink(tmp)
}

}
\seealso{
\code{\link{st_write}}, \code{\link{st_read}}
}
