% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streampress.R
\name{st_read_transpose}
\alias{st_read_transpose}
\title{Read Pre-Stored Transpose from a StreamPress File}
\usage{
st_read_transpose(path)
}
\arguments{
\item{path}{Path to a \code{.spz} file.}
}
\value{
A \code{dgCMatrix} representing CSC(A^T).
}
\description{
Read Pre-Stored Transpose from a StreamPress File
}
\seealso{
\code{\link{st_write}}, \code{\link{st_read}}
}
\keyword{internal}
