% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R, R/nmf_plots.R
\name{summary.nmf}
\alias{summary.nmf}
\alias{plot.nmfSummary}
\title{Summarize NMF factors}
\usage{
\method{summary}{nmf}(object, group_by, stat = "sum", ...)

\method{plot}{nmfSummary}(x, ...)
}
\arguments{
\item{object}{an object of class "\code{nmf}", usually, a result of a call to \code{\link{nmf}}}

\item{group_by}{a discrete factor giving groupings for samples or features. Must be of the same length as number of samples in \code{object$h} or number of features in \code{object$w}.}

\item{stat}{either \code{sum} (sum of factor weights falling within each group), or \code{mean} (mean factor weight falling within each group).}

\item{...}{arguments passed to or from other methods}

\item{x}{\code{nmfSummary} object, the result of calling \code{summary} on an \code{nmf} object}
}
\value{
\code{data.frame} with columns \code{group}, \code{factor}, and \code{stat}
}
\description{
\code{summary} method for class "\code{nmf}". Describes metadata representation in NMF factors. Returns object of class \code{nmfSummary}. Plot results using \code{plot}.
}
