% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tcrossprod}
\alias{tcrossprod}
\title{Cross-product of transpose}
\usage{
tcrossprod(x, y, encoding = "CSC")
}
\arguments{
\item{x}{sparse matrix object of class \code{dgCMatrix} of dimensions \code{m x n} with integral non-zero values}

\item{y}{dense matrix object of class \code{matrix} of dimensions \code{n x k}}

\item{encoding}{sparse encoding to be used}
}
\value{
dense matrix giving the transpose cross-product
}
\description{
Cross-product of transpose
}
\examples{
library(Matrix)
x <- rsparsematrix(100, 1000, density = 0.5,
       rand.x = function(n){sample(1 : 10, n, replace = TRUE)})
y <- matrix(runif(10 * 1000), 10, 1000)
plot(VSE::tcrossprod(x, y), t(Matrix::tcrossprod(x, y)))
}
