\name{timesTwo}
\alias{timesTwo}
\title{double, World!}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{An numberic value indicating if the development or
release version number should be returned, default is release.}
}
\value{
A \code{package_version} object with either the release
or development version.
}
\description{
Prints 'Hello, world!'.
}
\examples{
timesTwo(x = 20)
}