% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSE.R
\docType{package}
\name{VSE}
\alias{VSE}
\alias{VSE-package}
\title{VSE: Vectorized Sparse Encoding}
\description{
Proof-of-concept implementation of vectorized sparse
encodings for scalable and compressed structures for sparse data
}
\author{
Zach DeBruine
}
