% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aml.R
\docType{data}
\name{aml}
\alias{aml}
\title{Acute Myelogenous Leukemia cells}
\format{
dense matrix of features (DMRs) as rows and samples ("AML sample", "GMP", "LMPP", or "MEP") as columns.
}
\usage{
data(aml)
}
\description{
~800 differentially methylated regions (DMRs) in 123 Acute Myelogenous Leukemia (AML) samples, and 5 samples for putative cells of origin (GMP, LMPP, MEP) giving healthy DMR signatures.
}
\details{
AML methylation signatures differ from their cell of origin by additive methylation and subtractive methylation, in addition to tumor-specific heterogeneity. These AML tumors originated from one of three healthy cell types (GMP, LMPP, or MEP), the challenge is to classify the cell of origin based on these healthy cell type DMRs.
}
\keyword{datasets}
