% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartiteMatch.R
\name{bipartiteMatch}
\alias{bipartiteMatch}
\title{Bipartite graph matching}
\usage{
bipartiteMatch(x)
}
\arguments{
\item{x}{symmetric matrix giving the cost of every possible pairing}
}
\value{
List of "cost" and "pairs"
}
\description{
Hungarian algorithm for matching samples in a bipartite graph from a distance ("cost") matrix
}
\details{
This implementation was adapted from RcppHungarian, an Rcpp wrapper for the original C++ implementation by Cong Ma (2016).
}
