% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition.R
\name{bipartition}
\alias{bipartition}
\title{Bipartition a sample set}
\usage{
bipartition(
  A,
  tol = 1e-04,
  nonneg = TRUE,
  samples = NULL,
  seed = NULL,
  verbose = FALSE,
  calc_dist = FALSE,
  ...
)
}
\arguments{
\item{A}{dense or sparse matrix of features-by-samples}

\item{tol}{correlation distance (\eqn{1 - R^2}) between \eqn{w} across consecutive iterations at which to stop factorization}

\item{nonneg}{enforce non-negativity}

\item{samples}{samples to include in bipartition, numbered from 1 to \code{ncol(A)}. Default is \code{NULL} for all samples.}

\item{seed}{random seed for model initialization}

\item{verbose}{print model tolerances between iterations}

\item{calc_dist}{calculate the relative cosine distance of samples within a cluster to either cluster centroid. If \code{TRUE}, \code{calc_centers} will be set to \code{TRUE}.}

\item{...}{advanced parameters, see details}
}
\value{
A list giving the bipartition and useful statistics:
\itemize{
\item v       : vector giving difference between sample loadings between factors in a rank-2 factorization
\item dist    : relative cosine distance of samples within a cluster to centroids of assigned vs. not-assigned cluster
\item size1   : number of samples in first cluster (positive loadings in 'v')
\item size2   : number of samples in second cluster (negative loadings in 'v')
\item samples1: indices of samples in first cluster
\item samples2: indices of samples in second cluster
\item center1 : mean feature loadings across samples in first cluster
\item center2 : mean feature loadings across samples in second cluster
}
}
\description{
"Spectral" biparitioning by rank-2 matrix factorization
}
\details{
Spectral bipartitioning is a popular subroutine in divisive clustering. The sign of the difference between sample loadings in factors of a rank-2 matrix factorization
gives a bipartition that is nearly identical to an SVD.

Rank-2 matrix factorization by alternating least squares is faster than rank-2-truncated SVD (i.e. \emph{irlba}).

This function is a wrapper of \code{\link{nmf2}}, with additional calculations on the factorization model relevant to bipartitioning. See \code{\link{nmf2}} for details.
}
\section{Advanced parameters}{

The \code{...} argument hides several parameters that may be adjusted, although defaults should entirely satisfy:
\itemize{
\item \code{diag}, set to \code{FALSE} to disable model scaling by the diagonal. When \code{diag = FALSE}, symmetric factorization cannot occur and the difference between factors may not be linear with the second-left vector of an SVD.
\item \code{maxit}, default 100. Maximum number of alternating updates of \eqn{w} and \eqn{h}.
\item \code{calc_centers}, default TRUE. Calculate centroids for each cluster, giving mean feature loadings for all samples in each cluster. If \code{calc_centers = FALSE} and \code{calc_dist = TRUE}, then \code{calc_centers} will be set to \code{TRUE}.
}
}

\examples{
data(iris)
bipartition(iris[,1:4])

}
\references{
Kuang, D, Park, H. (2013). "Fast rank-2 nonnegative matrix factorization for hierarchical document clustering." Proc. 19th ACM SIGKDD intl. conf. on Knowledge discovery and data mining.
}
\seealso{
\code{\link{nmf2}}, \code{\link{dclust}}
}
\author{
Zach DeBruine
}
