% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartition.R
\name{bipartition}
\alias{bipartition}
\title{Bipartition a sample set}
\usage{
bipartition(
  A,
  tol = 1e-04,
  nonneg = TRUE,
  samples = NULL,
  seed = NULL,
  verbose = FALSE,
  calc_dist = FALSE,
  ...
)
}
\arguments{
\item{A}{matrix of features x samples, may be dense or sparse (a class of the "Matrix" package coercible to \code{Matrix::dgCMatrix})}

\item{tol}{correlation distance between \eqn{w} across consecutive iterations at which to stop factorization}

\item{nonneg}{apply non-negativity constraints}

\item{samples}{samples to include in bipartition, numbered from 1 to \code{ncol(A)}. Default is \code{NULL} for all samples.}

\item{seed}{random seed for initializing \eqn{h} with R RNG. May also \code{set.seed()} prior to calling this function.}

\item{verbose}{print NMF tolerances after each iteration to the console}

\item{calc_dist}{calculate the relative cosine distance of samples within a cluster to either cluster centroid. If \code{TRUE}, \code{calc_centers} will be set to \code{TRUE}.}

\item{...}{advanced parameters, see details}
}
\value{
A list giving the bipartition and useful statistics:
\itemize{
\item v       : vector giving difference between sample loadings between factors in a rank-2 factorization
\item dist    : relative cosine distance of samples within a cluster to centroids of assigned vs. not-assigned cluster
\item size1   : number of samples in first cluster (positive loadings in 'v')
\item size2   : number of samples in second cluster (negative loadings in 'v')
\item samples1: indices of samples in first cluster
\item samples2: indices of samples in second cluster
\item center1 : mean feature loadings across samples in first cluster
\item center2 : mean feature loadings across samples in second cluster
}
}
\description{
Spectral biparitioning by rank-2 matrix factorization
}
\details{
The difference between both factors in a rank-2 matrix factorization is approximately linear to the second-right singular vector, which is often used in spectral bipartitioning as a method of sample classification and clustering.

This function runs a rank-2 matrix factorization and returns the difference between sample loadings in the first and second factors (i.e. \eqn{h} matrix). Rank-2 matrix factorization by alternating least squares is faster than any current truncated SVD implementation (i.e. \eqn{irlba}).

There are backend specializations for dense or sparse input matrices. If a sparse matrix is provided, sparse optimizations will be used and generally are worthwhile if a matrix is >70\% sparse.
}
\section{Advanced parameters}{

\itemize{
\item \code{diag}, default TRUE. Enable model diagonalization to normalize factors to sum to 1, guarantee symmetry for symmetric factorizations, and achieve linearity of the bipartitioning vector with the second-left vector in rank-2 SVD.
\item \code{maxit}, default 100. Maximum number of alternating updates of \eqn{w} and \eqn{h}.
\item \code{calc_centers}, default TRUE. Calculate centroids for each cluster, giving mean feature loadings for all samples in each cluster. If \code{calc_centers = FALSE} and \code{calc_dist = TRUE}, then \code{calc_centers} will be set to \code{TRUE}.
}
}

\seealso{
\code{\link{nmf}}
}
\author{
Zach DeBruine
}
