% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidate.R, R/nmf_plots.R
\name{crossValidate}
\alias{crossValidate}
\alias{plot.nmfCrossValidate}
\title{Cross-validation for NMF}
\usage{
crossValidate(data, k, reps = 3, n = 0.05, verbose = FALSE, ...)

\method{plot}{nmfCrossValidate}(x, ...)
}
\arguments{
\item{data}{dense or sparse matrix of features in rows and samples in columns. Prefer \code{matrix} or \code{Matrix::dgCMatrix}, respectively}

\item{k}{array of factorization ranks to test}

\item{reps}{number of independent replicates to run}

\item{n}{fraction of values to handle as missing (default is 5\%, or \code{0.05})}

\item{verbose}{should updates be displayed when each factorization is completed}

\item{...}{parameters to \code{RcppML::nmf}, not including \code{data} or \code{k}}

\item{x}{\code{nmfCrossValidate} object, the result of \code{crossValidate}}
}
\value{
\code{data.frame} with class \code{nmfCrossValidate} with columns \code{rep}, \code{k}, and \code{value}
}
\description{
Find an "optimal" rank for a Non-Negative Matrix Factorization using cross-validation. Returns a \code{data.frame} with class \code{nmfCrossValidate}. Plot results using the \code{plot} class method.
}
\details{
A random speckled pattern of values is masked off during model fitting, and the mean squared error of prediction is evaluated after the model has reached the desired tolerance. The rank at which the model achieves the lowest error (best prediction accuracy) is the optimal rank.
}
\seealso{
\code{\link{nmf}}
}
