% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidate.R
\name{crossValidate}
\alias{crossValidate}
\title{Cross-validation for NMF}
\usage{
crossValidate(A, k, method = 1, reps = 5, verbose = 1, ...)
}
\arguments{
\item{A}{matrix of features-by-samples in dense or sparse format (preferred classes are \code{matrix} or \code{Matrix::dgCMatrix}, respectively). Prefer sparse storage when >50\% of values are zero.}

\item{k}{factorization ranks to test}

\item{method}{algorithm to use for cross-validation, see "details", either \code{1} or \code{2}.}

\item{reps}{number of independent replicates to run}

\item{verbose}{level of verbosity, \code{c(0, 1, 2)}}

\item{...}{parameters to \code{RcppML::nmf}, not including \code{A} or \code{k}}
}
\value{
\code{data.frame} with columns \code{rep}, \code{k}, and \code{value}, where \code{value} depends on the \code{method} selected (i.e. MSE, cost of bipartite matching)
}
\description{
Find an "optimal" rank for a Non-Negative Matrix Factorization using cross-validation.
}
\details{
Two algorithms are available for cross-validation and can be selected using the \code{method} parameter:
\enumerate{
\item \strong{Bi-cross-validation}: Data is split into an equally-sized 2x2 grid (A1, A2, B1, and B2). Data is trained on A1 and projected to A2 to B2 where the mean squared error of the model on B2 is measured.
\item \strong{Cost of bipartite matching} between independent replicates. Samples are partitioned into two sets and independently factorized, then correlation between factors in both \code{w} models is computed, factors are matched using bipartite matching on a cosine similarity cost matrix, and the mean cost of matched factors is returned.
}
}
\seealso{
\code{\link{nmf}}
}
