% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_net.R
\name{factor_add}
\alias{factor_add}
\title{Element-wise H addition (skip/residual connection)}
\usage{
factor_add(...)
}
\arguments{
\item{...}{Two or more \code{fn_node} objects with matching ranks.}
}
\value{
An \code{fn_node} of type "add".
}
\description{
Creates a node that adds H factors element-wise from multiple branches.
All branches must have the same rank k.
}
\examples{
data(aml)
inp <- factor_input(aml)
branch1 <- nmf_layer(inp, k = 5)
branch2 <- nmf_layer(inp, k = 5)
added <- factor_add(branch1, branch2)
}
\seealso{
\code{\link{factor_concat}}, \code{\link{factor_shared}}, \code{\link{factor_net}}
}
