% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_net.R
\name{factor_condition}
\alias{factor_condition}
\title{Concatenate conditioning metadata to a layer's H}
\usage{
factor_condition(input, Z)
}
\arguments{
\item{input}{An \code{fn_node} (typically an nmf_layer output).}

\item{Z}{Conditioning matrix (n x p) or (p x n). Will be oriented
to match H dimensions.}
}
\value{
An \code{fn_node} of type "condition".
}
\description{
Appends columns from a metadata matrix Z to the H factor before
passing downstream. The next layer's W learns to factor out the
conditioning variables.
}
\examples{
data(aml)
inp <- factor_input(aml)
layer1 <- nmf_layer(inp, k = 5)
# Condition on batch metadata (2 batches)
Z <- matrix(rep(c(1, 0, 0, 1), c(60, 75, 60, 75)), nrow = 135, ncol = 2)
conditioned <- factor_condition(layer1, Z)
}
