% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HawaiiBirds.R
\docType{data}
\name{HawaiiBirds}
\alias{HawaiiBirds}
\title{Bird species frequency in Hawaii}
\format{
list of two components: \code{counts} giving mean total counts of species in each grid, and \code{metadata} giving information about each grid (i.e. latitude, longitude, and island)
}
\usage{
data(HawaiiBirds)
}
\description{
Frequency of bird species observation within 1km-squared grids in Hawaii as recorded in the eBird project. Not all counts are complete because some grids are sampled better than others (i.e. urban areas, birding hotspots).
}
\details{
This dataset was obtained as follows:
\itemize{
\item All eBird observations in Hawaii were downloaded from the eBird website as of Oct. 2021
\item Only complete checklists were retained
\item Only non-X counts were retained
\item Species were removed with fewer than 50 observations or that were spuhs
\item Grids were defined by latitude/longitude rounded to two decimal places
\item Grids were removed with fewer than 10 checklists
\item Mean frequency of species in each grid was calculated based on the number of times a species was observed and the number of checklists submitted in that grid.
\item Grids were assigned to one of each major Hawaiian island based on geographical coordinates.
}
}
\keyword{datasets}
