% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{mse}
\alias{mse}
\title{Mean squared error of model fit}
\usage{
mse(object, data, ...)
}
\arguments{
\item{object}{fitted model}

\item{data}{data to which the model was fit}

\item{...}{additional arguments affecting the fit of the model}
}
\description{
\code{mse} is a generic function for evaluating the fit of models using Mean Squared Error (MSE).
}
\seealso{
\code{\link{mse.nmf}}
}
