% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\name{mse}
\alias{mse}
\title{Mean squared error of factor model}
\usage{
mse(w, d = NULL, h, data, mask = NULL, missing_only = FALSE, ...)
}
\arguments{
\item{w}{feature factor matrix (features as rows)}

\item{d}{scaling diagonal vector (if applicable)}

\item{h}{sample factor matrix (samples as columns)}

\item{data}{dense or sparse matrix of features in rows and samples in columns. Prefer \code{matrix} or \code{Matrix::dgCMatrix}, respectively}

\item{mask}{dense or sparse matrix of values in \code{data} to handle as missing. Prefer \code{Matrix::dgCMatrix}. Alternatively, specify "\code{zeros}" or "\code{NA}" to mask either all zeros or NA values.}

\item{missing_only}{only calculate mean squared error at masked values}

\item{...}{additional arguments}
}
\description{
Same as the \code{evaluate} S4 method for the \code{nmf} class, but allows one to input the `w`, `d`, `h`, and `data` independently.
}
