% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\name{mse}
\alias{mse}
\title{Mean squared error of factor model}
\usage{
mse(w, d = NULL, h, data, mask = NULL, missing_only = FALSE, ...)
}
\arguments{
\item{w}{feature factor matrix (features as rows)}

\item{d}{scaling diagonal vector (if applicable)}

\item{h}{sample factor matrix (samples as columns)}

\item{...}{additional arguments}
}
\description{
Same as the \code{evaluate} S4 method for the \code{nmf} class, but allows one to input the `w`, `d`, `h`, and `data` independently.
}
