% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Mean Squared Error (MSE) loss of a factor model}
\usage{
mse(A, w, d = NULL, h, precision = "float")
}
\arguments{
\item{A}{sparse matrix (of or coercible to \code{dgCMatrix}) of samples (columns) by features (rows)}

\item{w}{dense matrix of class \code{matrix} with factors (columns) by features (rows)}

\item{d}{vector (if applicable) giving giving coefficients of the diagonal}

\item{h}{dense matrix of class \code{matrix} with samples (columns) by factors (rows)}

\item{precision}{"float" or "double" precision}
}
\value{
mean squared error of the factorization
}
\description{
Mean squared error of factor models \eqn{W} and \eqn{H} given sparse matrix \eqn{A}
}
