% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_metrics.R
\name{print.fn_classifier_eval}
\alias{print.fn_classifier_eval}
\title{Print a classifier evaluation result}
\usage{
\method{print}{fn_classifier_eval}(x, ...)
}
\arguments{
\item{x}{An \code{fn_classifier_eval} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a classifier evaluation result
}
