% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_nmf.r
\name{project}
\alias{project}
\title{Project a model onto new data}
\usage{
project(w, data, L1 = 0, L2 = 0, mask = NULL, upper_bound = 0, ...)
}
\arguments{
\item{w}{matrix of features (rows) by factors (columns), corresponding to rows in \code{data}}

\item{data}{a dense or sparse matrix}

\item{L1}{L1/LASSO penalty}

\item{L2}{L2/Ridge penalty}

\item{mask}{masking on data values}

\item{upper_bound}{maximum value permitted in least squares solutions, essentially a bounded-variable least squares problem between 0 and \code{upper_bound}}

\item{...}{arguments passed to \code{predict.nmf}}
}
\description{
Equivalent to \code{predict} method for NMF, but requires only the \code{w} matrix to be supplied and not the entire NMF model. Use NNLS to project a basis factor model onto new samples.
}
\details{
See \code{\link{nmf}} for more info, as well as the \code{predict} method for NMF.
}
