% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_sp_read}
\alias{Rcpp_sp_read}
\title{Read a SparsePress (.spz) file into a dgCMatrix}
\usage{
Rcpp_sp_read(path, cols = NULL, reorder = TRUE)
}
\arguments{
\item{path}{Input file path (.spz)}

\item{cols}{Optional integer vector of column indices to read (1-indexed)}

\item{reorder}{Whether to undo row permutation (default TRUE)}
}
\value{
A sparse matrix (dgCMatrix)
}
\description{
Read a SparsePress (.spz) file into a dgCMatrix
}
\keyword{internal}
