% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppML-package.R
\docType{package}
\name{RcppML}
\alias{RcppML}
\alias{_PACKAGE}
\alias{RcppML-package}
\title{RcppML: Rcpp Machine Learning Library}
\description{
High-performance non-negative matrix factorization and linear model projection for sparse matrices, and fast non-negative least squares implementations
}
\author{
Zach DeBruine
}
