% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateNMF.R
\name{simulateNMF}
\alias{simulateNMF}
\title{Simulate an NMF dataset}
\usage{
simulateNMF(nrow, ncol, k, noise = 0.5, dropout = 0.5, seed = NULL)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{k}{true rank of simulated model}

\item{noise}{standard deviation of Gaussian noise centered at 0 to add to input matrix. Any negative values after noise addition are set to 0.}

\item{dropout}{density of dropout events}

\item{seed}{seed for random number generation}
}
\value{
dense matrix
}
\description{
Generate a random matrix that follows some defined NMF model to test NMF factorizations. Adapts methods from \code{NMF::syntheticNMF}.
}
