% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_methods.R
\name{sparsity}
\alias{sparsity}
\alias{sparsity.nmf}
\title{Calculate sparsity of an object}
\usage{
sparsity(object, ...)

\method{sparsity}{nmf}(object, ...)
}
\arguments{
\item{object}{object to be evaluated}

\item{...}{arguments passed to or from other methods}
}
\description{
\code{sparsity} is a generic function for calculating the (marginal) sparsity of a model.
}
\details{
For \code{\link{nmf}} models, the sparsity of each factor is computed and summarized for \eqn{w} and \eqn{h} matrices. A long \code{data.frame} with columns \code{factor}, \code{sparsity}, and \code{model} is returned.
}
