% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streampress.R
\name{st_add_transpose}
\alias{st_add_transpose}
\title{Add Transpose Section to an Existing StreamPress File}
\usage{
st_add_transpose(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to a \code{.spz} file.}

\item{verbose}{Logical; print progress. Default TRUE.}
}
\value{
Invisibly returns the path.
}
\description{
Add Transpose Section to an Existing StreamPress File
}
